+++ tests/test_utils/mod.rs
@@ -1,6 +1,5 @@
-use envisim_samplr::{SampleOptions, Sampler};
+use envisim_samplr::{SampleOptions, Sampler, SamplingError};
 use envisim_utils::utils::sum;
-pub use envisim_utils::SamplingError;
 use rand::Rng;
 
 #[allow(dead_code)]